// src/api/models/StreamVodsInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jingle Jam Community Page Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StreamVodsInner
 */
export interface StreamVodsInner {
    /**
     * 
     * @type {string}
     * @memberof StreamVodsInner
     */
    label?: string;
    /**
     * 
     * @type {string}
     * @memberof StreamVodsInner
     */
    link: string;
}

/**
 * Check if a given object implements the StreamVodsInner interface.
 */
export function instanceOfStreamVodsInner(value: object): value is StreamVodsInner {
    if (!('link' in value) || value['link'] === undefined) return false;
    return true;
}

export function StreamVodsInnerFromJSON(json: any): StreamVodsInner {
    return StreamVodsInnerFromJSONTyped(json, false);
}

export function StreamVodsInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): StreamVodsInner {
    if (json == null) {
        return json;
    }
    return {
        
        'label': json['label'] == null ? undefined : json['label'],
        'link': json['link'],
    };
}

export function StreamVodsInnerToJSON(value?: StreamVodsInner | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'label': value['label'],
        'link': value['link'],
    };
}